% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/todotxt.R
\docType{package}
\name{todotxt}
\alias{todotxt}
\alias{todotxt-package}
\alias{todotxt}
\title{todotxt}
\usage{
todotxt(todos)
}
\arguments{
\item{todos}{Path to todo.txt file}
}
\value{
Todo list as a tibble.
}
\description{
todotxt

Parse todo.txt files
}
